refactor: extract tui file surface
