+++ scripts/rust-source-size-baseline.txt
@@ -5,4 +5,4 @@ crates/shell-core/src/control.rs 1183
 crates/shell-core/src/model.rs 1320
 crates/shell-core/src/reducer.rs 2723
 crates/shell-core/src/terminal_profile.rs 1852
-crates/tui/src/file_backed.rs 2753
+crates/tui/src/file_backed.rs 1991
